+++ main.c
@@ -6,21 +6,6 @@
 #include <stdlib.h>
 #include "structure.h"
 
-int is_ifs(char c)
-{
-	if (c == ' ' || c == '\t' || c == '\n')
-		return (1);
-	return (0);
-}
-
-int	is_metacharacter(char c)
-{
-	if (c == '|' || c == '&' || c == '(' || c ==')'
-	|| c == '>' || c == '<')
-		return (1);
-	return (0);
-}
-
 void print_node(t_token *node)
 {
 	if (!node)
@@ -29,101 +14,6 @@ void print_node(t_token *node)
 	printf("token : %d\n", node->token);
 }
 
-void del_token_node(t_token *node)
-{
-	if (!node)
-		return ;
-	free(node->str);
-	free(node);
-}
-
-void	ft_tokenadd(t_list *head, char *str, int tokennum)
-{
-		t_list	*new_node;
-		t_token	*content;
-
-		content = ft_tokennew(str, tokennum);
-		if (!content)
-			ft_lstclear(&head, (void *)del_token_node);
-		new_node = ft_lstnew(content);
-		ft_lstadd_back(&head, new_node);
-}
-
-int ft_token_num(char *line, int i)
-{
-	if (line[i] == '<' && line[i + 1] == '<')
-		return (REDIRECT_HEREDOC);
-	else if (line[i] == '<')
-		return (REDIRECT_IN);
-	else if (line[i] == '>' && line[i + 1] == '>')
-		return (REDIRECT_APPEND);
-	else if (line[i] == '>')
-		return (REDIRECT_OUT);
-	else if (line[i] == '|' && line[i + 1] == '|')
-		return (OR_OPERATOR);
-	else if (line[i] == '|')
-		return (PIPE);
-	else if (line[i] == '&' && line[i + 1] == '&')
-		return (AND_OPERATOR);
-	else if (line[i] == '(')
-		return (L_PAREN);
-	else if (line[i] ==')')
-		return (R_PAREN);
-	else
-		return (0);
-		// ㅇㅣ ㅂㅜ부ㄴ 에러처리 어떻게 할지 생각해보기
-}
-
-void ft_token_metachar(char *line, int *i, int start, t_list *head)
-{
-	char	*str;
-	int		tokennum;
-
-	tokennum = ft_token_num(line, *i);
-	if (tokennum == ERROR)
-	{
-		printf("Syntax Error\n"); // error 문구 생각하기
-		exit(-1); // exit 코드 생각
-		//free 해줘야하는지 고려하기..
-	}
-	else if (tokennum ==  REDIRECT_IN || tokennum == REDIRECT_OUT || tokennum == L_PAREN 
-	|| tokennum == R_PAREN || tokennum == PIPE)
-	 	*i += 1;
-	else if (tokennum == REDIRECT_APPEND || tokennum == REDIRECT_HEREDOC
-	|| tokennum == AND_OPERATOR || tokennum == OR_OPERATOR)
-		*i += 2;
-	str = ft_substr(line, start, *i - start);
-	ft_tokenadd(head,str,tokennum);
-}
-
-t_list *tokenizer(char *line, t_list *head)
-{
-	int		i;
-	int		start;
-	char	*str;
-
-	i = 0;
-	while (line[i])
-	{
-		while (line[i] && is_ifs(line[i]) && !is_metacharacter(line[i]))
-			i++;
-		start = i;
-		while (line[i] && !is_ifs(line[i]) && !is_metacharacter(line[i]))
-			i++;
-		if (i - start > 0)
-		{
-			str = ft_substr(line, start, i - start);
-			ft_tokenadd(head,str,WORD);
-		}
-		else if (is_metacharacter(line[i]))
-		{
-			ft_token_metachar(line, &i, start, head);
-		}
-		// 이전것까지 토큰리스트에 추가
-		// is_metacharacter인지 확인하고 metacharacter인 경우 token_list에 추가
-	}
-	return head;
-}
 
 void leaks (void){
 	system("leaks minishell");
@@ -141,8 +31,7 @@ int main(void)
 		line = readline("examshell : ");
 		add_history(line); 
 		tokenizer(line, head);
-		printf("%p\n",head);
 		ft_lstiter(head, (void *)print_node);
-		ft_lstclear(&head, (void *)del_token_node);
+		ft_lstclear(&head, (void *)ft_del_token_node);
 	}
 }
